requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pypdf2>=3.0.0
python-docx>=0.8.11
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# For local LLM
llama-cpp-python>=0.2.0
# OR ollama python package if using Ollama
# ollama>=0.1.0

# For structured output
langchain>=0.1.0
pydantic>=2.0.0

# Development
jupyter>=1.0.0
ipython>=8.0.0